create and incorporate teachingstaff class
