+++ highfive.rb
@@ -1,0 +1,6 @@
+
+module Highfiveable
+	def offer_high_five
+    "High five!"
+  end
+end
